app/requirements.txt
Flask==1.0.2
Flask-PyMongo==2.3.0
requests==2.20.1
urllib3==1.22
python-dotenv==0.19.0
pymongo[bson]
pytz==2022.7.1